aboutMe/requirements.txt
asgiref==3.4.1
django-crispy-forms==1.13.0
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
